This code demonstrates a partially functional Windows service that does a couple of critical things:
   1. Authenticates as a user and creates a file in that user's home
      directory.
   2. Opens a window on the Winlogon desktop.

Unfortunately, #2 isn't quite working yet. It works correctly for the
Application desktop (even though it is specifically changing the process
desktop to the Winlogon desktop). The problem that I am currently trying
to overcome is that interactive services are not started until after a
user logs in. Non-interactive services, OTOH, are started at boot time
without requiring a user to be logged in.