When a UserPerspective is first handled credentials, use them to perform the Windows logon operation so that we grant window opening permissions to the authenticated user right away. At the time of logout (thanks, Twisted, for making that piece easy!), clean up authentication pieces and remove the window opening permissions. If all goes well, we shouldn't have to worry at all about filling up the ACL for the window station or the desktop.
As a side note, I think that I made a change somewhere in
util/process.py that had broken the ability for text written to
stdout and stderr to be piped back to the client GUI. When I was
cleaning things up for this commit, I seem to undone my mistake, so
now everything related to launching is working on Windows. The next
thing to try will getting drive letter mappings to work automatically.