When running on an OS that uses X11 (currently determined using the faulty assumption that a non-Windows OS runs X11), grant the authenticated user permission to open windows on the display. This works when a user is logged in and when not. This elminiates the need for external scripts or X server modifications that would otherwise do this same thing.
This change fixes the problem with X11 applications not launching when
maestrod is run as root. Fortunately, Dan gave me lots of useful information
in getting this implemented and in fixing bugs with what I had written.

Note that additional settings have been introduced to the maestrod
configuration file. These are used to account for the fact that xauth(1)
is not always in /usr/X11R6/bin and that GDM is not always the display
manager.