To facilitate the asynchronous handling of output that gets piped back to the client side, set things up so that stdout and stderr in the client process both write into the same pipe. This results in the same ordering of output that would be seen on the console when executing the process outside the context of Maestro, and it keeps us from having to worry about handling stderr and stdout separately in the Launch Service.
This change is intentionally small. I did not want to risk complicating
a future import of process.py from its vendor branch if lots of changes
should occur to the child process spawning code. Because of this, more
work is currently done when creating pipes and whatnot than is actually
necessary given that only two pipes are needed instaed of three.
Therefore, this could be made more effecient if necessary.