When performing a remove operation on a collection of multiple roots, we have to be able to address those roots so that any one of them can be removed. To that end, I have extended the code for handling the remove operation so that the path in the <remove> element can start with the list index syntatic structure (e.g., <remove id="@3"/>).
The basic problem here is that the ID used with a remove command is a
relative path, but using "." is not specific enough when there are multiple
roots returned through dereferencing. Secondly, a string identifier in this
context refers to a *child* of the returned root(s). Hence, indexing
numerically is the only way that we have to refer to a root when removing
without having to repeat the root name in every usage of the remove command.

Aron's objection to this approach is noted.