Rewrote the builders that use pyuic4 and pyrcc4. My primary goal here is to fix a long-standing problem wherein SCons would not run pyrcc4 if a file referenced by a .qrc file changed. Now, the dependencies of .qrc files are known to SCons.
Another change is that the builder that runs pyrcc4 generates a .py file
based on what is expected by the code generated for pyuic4. When pyuic4 sees
a reference to filename.qrc reference in a .ui file, it adds a line to read
the contents of the compiled resource file ('import filename_rc'). The
builder previously generated filename.py from filename.qrc, and that meant
that the code generated by pyuic4 was invalid. This is now fixed. However,
there still remains a problem where relative paths to .qrc files result in
Python import statements being generated that also use a relative path. I do
not yet know how to address this problem.